simplified_pipeline/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-experimental>=0.0.47
langchain-core>=0.1.0

# LLM providers (install the ones you need)
langchain-openai>=0.0.5  # For OpenAI models
# langchain-anthropic>=0.0.1  # For Anthropic models
# langchain-google-genai>=0.0.1  # For Google models

# PDF processing
pymupdf>=1.23.0
chardet>=5.0.0

# Neo4j
neo4j>=5.0.0

# Utilities
python-dotenv>=1.0.0

# Optional: For embeddings and vector search
# openai>=1.0.0
# tiktoken>=0.5.0